+++ src/main/java/com/rettichlp/UnicacityAddon/commands/faction/FactionInfoCommand.java
@@ -7,7 +7,7 @@ import com.rettichlp.UnicacityAddon.base.faction.Faction;
 import com.rettichlp.UnicacityAddon.base.registry.annotation.UCCommand;
 import com.rettichlp.UnicacityAddon.base.text.ColorCode;
 import com.rettichlp.UnicacityAddon.base.text.Message;
-import com.rettichlp.UnicacityAddon.events.FactionInfoEventHandler;
+import com.rettichlp.UnicacityAddon.events.faction.FactionInfoEventHandler;
 import net.minecraft.command.ICommand;
 import net.minecraft.command.ICommandSender;
 import net.minecraft.server.MinecraftServer;
