+++ src/main/java/com/rettichlp/UnicacityAddon/commands/faction/badfaction/GaggedCommand.java
@@ -4,7 +4,7 @@ import com.rettichlp.UnicacityAddon.base.abstraction.AbstractionLayer;
 import com.rettichlp.UnicacityAddon.base.abstraction.UPlayer;
 import com.rettichlp.UnicacityAddon.base.registry.annotation.UCCommand;
 import com.rettichlp.UnicacityAddon.base.utils.ForgeUtils;
-import com.rettichlp.UnicacityAddon.events.GaggedEventHandler;
+import com.rettichlp.UnicacityAddon.events.faction.badfaction.GaggedEventHandler;
 import net.minecraft.command.ICommand;
 import net.minecraft.command.ICommandSender;
 import net.minecraft.server.MinecraftServer;
